refactor: split input event handlers into view_input_handlers.go
The ReadOnly additions pushed gui/view_input.go to 826 lines, over the
800-line gate (scripts/large-files.sh, make large-files). CI caught it;
the local run did not, because `make check` was not part of the
pre-push loop.

Move the event-handler layer — inputTextChange, the OnChar/OnKeyDown/
OnKeyUp closure constructors, inputKeyMutatesText, and the per-key
helpers — to view_input_handlers.go. Pure code motion: no logic
changes, and the moved block needs no imports (log and time are used
only by the remainder).

view_input.go 826 -> 546, view_input_handlers.go 287.
